+++ tapis-notificationslib/src/main/java/edu/utexas/tacc/tapis/notifications/dao/NotificationsDao.java
@@ -24,7 +24,8 @@ public interface NotificationsDao
 
   void migrateDB() throws TapisException;
 
-  int getNextSeriesSeqCount(ResourceRequestUser rUser, String seriesId) throws TapisException;
+  int getNextSeriesSeqCount(ResourceRequestUser rUser, String tenant, String source, String subject, String seriesId)
+          throws TapisException;
 
   // -----------------------------------------------------------------------
   // ------------------------- Subscriptions -------------------------------