+++ tapis-notificationslib/src/main/resources/db/migration/V002__addSeriesSeqCount.sql
@@ -1,7 +1,11 @@
 -- Table for tracking sequence IDs associated with a series
 CREATE TABLE IF NOT EXISTS series_seq_count
 (
-    id TEXT PRIMARY KEY,
-    seq_count INTEGER NOT NULL DEFAULT 0
+    tenant TEXT NOT NULL,
+    source TEXT NOT NULL,
+    subject TEXT NOT NULL,
+    seriesId TEXT NOT NULL,
+    seq_count INTEGER NOT NULL DEFAULT 0,
+    PRIMARY KEY (tenant, source, subject, seriesId)
 );
 ALTER TABLE series_seq_count OWNER TO tapis_ntf;